+++ src/OpenTelemetry.Instrumentation.SqlClient/SqlClientTraceInstrumentationOptions.cs
@@ -1,13 +1,11 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System.Data;
 using System.Diagnostics;
 using Microsoft.Extensions.Configuration;
 #if NET
 using OpenTelemetry.Instrumentation.SqlClient.Implementation;
 #endif
-using OpenTelemetry.Trace;
 using static OpenTelemetry.Internal.DatabaseSemanticConventionHelper;
 
 namespace OpenTelemetry.Instrumentation.SqlClient;
@@ -21,6 +19,7 @@ namespace OpenTelemetry.Instrumentation.SqlClient;
 public class SqlClientTraceInstrumentationOptions
 {
     internal const string ContextPropagationLevelEnvVar = "OTEL_DOTNET_EXPERIMENTAL_SQLCLIENT_ENABLE_TRACE_CONTEXT_PROPAGATION";
+    internal const string SetDbQueryParametersEnvVar = "OTEL_DOTNET_EXPERIMENTAL_SQLCLIENT_ENABLE_TRACE_DB_QUERY_PARAMETERS";
 
     /// <summary>
     /// Initializes a new instance of the <see cref="SqlClientTraceInstrumentationOptions"/> class.
@@ -46,46 +45,23 @@ public class SqlClientTraceInstrumentationOptions
         {
             this.EnableTraceContextPropagation = enableTraceContextPropagation;
         }
+
+        if (configuration!.TryGetBoolValue(
+                SqlClientInstrumentationEventSource.Log,
+                SetDbQueryParametersEnvVar,
+                out var setDbQueryParameters))
+        {
+            this.SetDbQueryParameters = setDbQueryParameters;
+        }
 #endif
     }
-
-    /// <summary>
-    /// Gets or sets a value indicating whether or not the <see cref="SqlClientInstrumentation"/>
-    /// should add the text of commands as the <see cref="SemanticConventions.AttributeDbStatement"/> tag.
-    /// Default value: <see langword="false"/>.
-    /// </summary>
-    /// <remarks>
-    /// <para>
-    /// <b>WARNING: SetDbStatementForText will capture the raw
-    /// <c>CommandText</c>. Make sure your <c>CommandText</c> property never
-    /// contains any sensitive data.</b>
-    /// </para>
-    /// <para><b>SetDbStatementForText is supported on all runtimes.</b>
-    /// <list type="bullet">
-    /// <item>On .NET and .NET Core SetDbStatementForText only applies to
-    /// <c>SqlCommand</c>s with <see cref="CommandType.Text"/>.</item>
-    /// <item>On .NET Framework SetDbStatementForText applies to all
-    /// <c>SqlCommand</c>s regardless of <see cref="CommandType"/>.
-    /// <list type="bullet">
-    /// <item>When using <c>System.Data.SqlClient</c> use
-    /// SetDbStatementForText to capture StoredProcedure command
-    /// names.</item>
-    /// <item>When using <c>Microsoft.Data.SqlClient</c> use
-    /// SetDbStatementForText to capture Text, StoredProcedure, and all
-    /// other command text.</item>
-    /// </list></item>
-    /// </list>
-    /// </para>
-    /// </remarks>
-    public bool SetDbStatementForText { get; set; }
 
     /// <summary>
     /// Gets or sets an action to enrich an <see cref="Activity"/> with the
     /// raw <c>SqlCommand</c> object.
     /// </summary>
     /// <remarks>
-    /// <para><b>Enrich is only executed on .NET and .NET Core
-    /// runtimes.</b></para>
+    /// <para><b>Enrich is only executed on .NET runtimes.</b></para>
     /// The parameters passed to the enrich action are:
     /// <list type="number">
     /// <item>The <see cref="Activity"/> being enriched.</item>
@@ -103,8 +79,7 @@ public class SqlClientTraceInstrumentationOptions
     /// collect telemetry about a command.
     /// </summary>
     /// <remarks>
-    /// <para><b>Filter is only executed on .NET and .NET Core
-    /// runtimes.</b></para>
+    /// <para><b>Filter is only executed on .NET runtimes.</b></para>
     /// Notes:
     /// <list type="bullet">
     /// <item>The first parameter passed to the filter function is the raw
@@ -126,13 +101,13 @@ public class SqlClientTraceInstrumentationOptions
     /// langword="false"/>.
     /// </summary>
     /// <remarks>
-    /// <para><b>RecordException is only supported on .NET and .NET Core
-    /// runtimes.</b></para>
+    /// <para><b>RecordException is only supported on .NET runtimes.</b></para>
     /// <para>For specification details see: <see
     /// href="https://github.com/open-telemetry/semantic-conventions/blob/main/docs/exceptions/exceptions-spans.md"/>.</para>
     /// </remarks>
     public bool RecordException { get; set; }
 
+#if !NETFRAMEWORK
     /// <summary>
     /// Gets or sets a value indicating whether or not the <see cref="SqlClientInstrumentation"/>
     /// should add the names and values of query parameters as the <c>db.query.parameter.{key}</c> tag.
@@ -145,10 +120,11 @@ public class SqlClientTraceInstrumentationOptions
     /// contain any sensitive data.</b>
     /// </para>
     /// <para>
-    /// <b>SetDbQueryParameters is only supported on .NET and .NET Core runtimes.</b>
+    /// <b>SetDbQueryParameters is only supported on .NET runtimes.</b>
     /// </para>
     /// </remarks>
-    public bool SetDbQueryParameters { get; set; }
+    internal bool SetDbQueryParameters { get; set; }
+#endif
 
     /// <summary>
     /// Gets or sets a value indicating whether the old database attributes should be emitted.