+++ test/OpenTelemetry.OpAmp.Client.Tests/PlainHttpTransportTest.cs
@@ -1,8 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using Google.Protobuf;
-using OpAmp.Proto.V1;
 using OpenTelemetry.OpAmp.Client.Internal;
 using OpenTelemetry.OpAmp.Client.Internal.Transport.Http;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
@@ -13,33 +11,35 @@ namespace OpenTelemetry.OpAmp.Client.Tests;
 
 public class PlainHttpTransportTest
 {
-    [Fact]
-    public async Task PlainHttpTransport_SendReceiveCommunication()
+    [Theory]
+    [InlineData(true)]
+    [InlineData(false)]
+    public async Task PlainHttpTransport_SendReceiveCommunication(bool useSmallPackets)
     {
         // Arrange
-        using var opAmpServer = new OpAmpFakeHttpServer();
+        using var opAmpServer = new OpAmpFakeHttpServer(useSmallPackets);
         var opAmpEndpoint = opAmpServer.Endpoint;
 
-        var mockListener = new MockListener();
+        using var mockListener = new MockListener();
         var frameProcessor = new FrameProcessor();
         frameProcessor.Subscribe(mockListener);
 
         var httpTransport = new PlainHttpTransport(opAmpEndpoint, frameProcessor);
 
-        var uid = ByteString.CopyFrom(Guid.NewGuid().ToByteArray());
-        var frame = new AgentToServer() { InstanceUid = uid };
+        var mockFrame = FrameGenerator.GenerateMockAgentFrame(useSmallPackets);
 
         // Act
-        await httpTransport.SendAsync(frame, CancellationToken.None);
+        await httpTransport.SendAsync(mockFrame.Frame, CancellationToken.None);
 
         // Assert
         var serverReceivedFrames = opAmpServer.GetFrames();
         var clientReceivedFrames = mockListener.Messages;
+        var receivedTextData = clientReceivedFrames.First().CustomMessage.Data.ToStringUtf8();
 
         Assert.Single(serverReceivedFrames);
-        Assert.Equal(uid, serverReceivedFrames.First().InstanceUid);
+        Assert.Equal(mockFrame.Uid, serverReceivedFrames.First().InstanceUid);
 
         Assert.Single(clientReceivedFrames);
-        Assert.Equal("Response from OpAmpFakeHttpServer", clientReceivedFrames.First().CustomMessage.Data.ToStringUtf8());
+        Assert.StartsWith("This is a mock server frame for testing purposes.", receivedTextData);
     }
 }