+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -6,3 +6,10 @@
   ([#2917](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2917))
 * Added support for OpAMP Plain HTTP transport.
   ([#2926](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2926))
+* Added support for OpAMP WebSocket transport.
+  ([#3064](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3064),
+  [#3092](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3092))
+* Added support for heartbeat messages.
+  ([#3095](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3095))
+* Added EventSource `OpenTelemetry-OpAmp-Client` for diagnostics.
+  ([#3102](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3102))